refactor(phase-4): decompose god objects into focused sub-modules
Decompose the two largest files in the codebase into focused, single-responsibility
sub-modules using the mixin pattern:

golf_launcher.py (2557 -> 684 LOC):
- launcher_constants.py: Shared constants, lazy imports, feature flags
- launcher_ui_setup.py: Menu bar, top bar, grid area, bottom bar, search, console
- launcher_theme.py: Theme application, theme menus, plot theme
- launcher_simulation.py: Simulation launching, dependency checking
- launcher_dialogs.py: Dialogs, settings, keyboard shortcuts, toast

statistical_analysis.py (2236 -> 151 LOC):
- coordination_metrics.py: Coupling angles, CRP, coordination
- nonlinear_dynamics.py: Lyapunov, RQA, entropy, fractal analysis
- pca_analysis.py: PCA, principal movements, kinematic sequence
- power_work_metrics.py: Work, power, impulse, stiffness
- reporting.py: Reports, CSV export, frequency, jerk, swing profile

All files stay under 1000 LOC. Backward compatibility maintained through
composition class inheritance and __all__ re-exports.

Closes #1263
