+++ src/engines/loaders.py
@@ -27,7 +27,6 @@ def load_mujoco_engine(suite_root: Path) -> PhysicsEngine:
         from src.engines.physics_engines.mujoco.python.mujoco_humanoid_golf.physics_engine import (
             MuJoCoPhysicsEngine,
         )
-
         from src.shared.python.engine_probes import MuJoCoProbe
 
         probe = MuJoCoProbe(suite_root)
@@ -72,7 +71,6 @@ def load_drake_engine(suite_root: Path) -> PhysicsEngine:
         from src.engines.physics_engines.drake.python.drake_physics_engine import (
             DrakePhysicsEngine,
         )
-
         from src.shared.python.engine_probes import DrakeProbe
 
         probe = DrakeProbe(suite_root)
@@ -123,7 +121,6 @@ def load_pinocchio_engine(suite_root: Path) -> PhysicsEngine:
         from src.engines.physics_engines.pinocchio.python.pinocchio_physics_engine import (
             PinocchioPhysicsEngine,
         )
-
         from src.shared.python.engine_probes import PinocchioProbe
 
         probe = PinocchioProbe(suite_root)
@@ -165,7 +162,6 @@ def load_opensim_engine(suite_root: Path) -> PhysicsEngine:
         from src.engines.physics_engines.opensim.python.opensim_physics_engine import (
             OpenSimPhysicsEngine,
         )
-
         from src.shared.python.engine_probes import OpenSimProbe
 
         probe = OpenSimProbe(suite_root)
@@ -190,7 +186,6 @@ def load_myosim_engine(suite_root: Path) -> PhysicsEngine:
         from src.engines.physics_engines.myosuite.python.myosuite_physics_engine import (
             MyoSuitePhysicsEngine,
         )
-
         from src.shared.python.engine_probes import MyoSimProbe
 
         probe = MyoSimProbe(suite_root)
